Fixed lighting bugs, self-intersection, refactored

+++ include/color.h
@@ -19,8 +19,7 @@ constexpr Color operator+(const Color &a, const Color &b) noexcept {
 }
 
 constexpr Color operator*(const float c, const Color &a) noexcept {
-  return Color{static_cast<float>(c * a.r), static_cast<float>(c * a.g),
-               static_cast<float>(c * a.b)};
+  return Color{c * a.r, c * a.g, c * a.b};
 }
 
 constexpr Color operator*(const double c, const Color &a) noexcept {
